# Conflicts:
#	projects/lib-commons/src/main/java/com/mbrlabs/mundus/commons/loader/AssimpWorker.java

+++ projects/lib-core/src/main/java/net/nevinsky/abyssus/core/ModelBuilder.java
@@ -61,7 +61,7 @@ public class ModelBuilder {
 
     private MeshBuilder getBuilder(final VertexAttributes attributes) {
         for (final MeshBuilder mb : builders) {
-            if (mb.getAttributes().equals(attributes) && mb.lastIndex() < Short.MAX_VALUE / 2) {
+            if (mb.getAttributes().equals(attributes) && mb.lastIndex() < Integer.MAX_VALUE / 2) {
                 return mb;
             }
         }
